move input module to its own crate, move output mod
