+++ crates/bindings-macro-input/src/sats.rs
