+++ crates/bindings-macro-input/src/sym.rs
