+++ crates/bindings-macro-input/src/table.rs
