+++ crates/bindings-macro-input/src/util.rs
