--- crates/bindings-macro/src/input.rs
@@ -1,4 +1,0 @@
-pub mod table;
-pub mod util;
-pub mod sats;
-pub mod sym;
