+++ crates/bindings-macro/src/lib.rs
@@ -8,10 +8,9 @@
 //
 // (private documentation for the macro authors is totally fine here and you SHOULD write that!)
 
-mod input;
-mod output;
 mod reducer;
 mod sats;
+mod table;
 mod util;
 
 use proc_macro::TokenStream as StdTokenStream;
