--- crates/bindings-macro/src/output.rs
@@ -1,1 +1,0 @@
-pub mod table;
