+++ crates/bindings-macro/src/table.rs
