+++ docker-compose-release.yml
@@ -12,6 +12,7 @@ services:
       - ./crates/lib:/usr/src/app/crates/lib
       - ./crates/sats:/usr/src/app/crates/sats
       - ./crates/bindings:/usr/src/app/crates/bindings
+      - ./crates/bindings-macro-input:/usr/src/app/crates/bindings-macro-input
       - ./crates/bindings-macro:/usr/src/app/crates/bindings-macro
       - ./crates/bindings-sys:/usr/src/app/crates/bindings-sys
       - ./crates/vm:/usr/src/app/crates/vm
